Build telemetry-all from sourceSets rather than project dependencies
Instead of trying to manipulate the various subproject JARs into
a single JAR that could be a module, this approach collects the
source directories from those subprojects and creates a special
source set within the project. in reality everything gets built again,
but now everything should be within a single toplevel package.